+++ tests/unit/no-top-level-side-effects.test.ts
@@ -9,14 +9,24 @@ const valid: RuleTester.ValidTestCase[] = [
       (function() {
         return '';
       })();
-    `
+    `,
+    options: [
+      {
+        allowIIFE: true
+      }
+    ]
   },
   {
     code: `
       (() => {
         return '';
       })();
-    `
+    `,
+    options: [
+      {
+        allowIIFE: true
+      }
+    ]
   },
   {
     code: `
@@ -164,11 +174,6 @@ const invalid: RuleTester.InvalidTestCase[] = [
         return '';
       })();
     `,
-    options: [
-      {
-        allowIIFE: false
-      }
-    ],
     errors
   },
   {
@@ -177,9 +182,15 @@ const invalid: RuleTester.InvalidTestCase[] = [
         return '';
       })();
     `,
+    errors
+  },
+  {
+    code: `
+      console.log('hello world');
+    `,
     options: [
       {
-        allowIIFE: false
+        allowIIFE: true
       }
     ],
     errors