+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Versioning].
 
 ## [Unreleased]
 
+- (`9ddf3fe`) _(Breaking)_ Rename the rule `no-top-level-side-effect` to
+  `no-top-level-side-effects`.
 - (`92a3f58`) Improve performance of `no-top-level-variables`.
 
 ## [0.3.0] - 2023-01-13