+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning].
 
 - (`9ddf3fe`) _(Breaking)_ Rename the rule `no-top-level-side-effect` to
   `no-top-level-side-effects`.
+- (`cec69d5`) _(Breaking)_ Report top-level IIFEs by default.
 - (`92a3f58`) Improve performance of `no-top-level-variables`.
 
 ## [0.3.0] - 2023-01-13