+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Versioning].
 
 ## [Unreleased]
 
+- _No changes yet_
+
+## [1.0.0] - 2023-02-25
+
 - (`9ddf3fe`) _(Breaking)_ Rename the rule `no-top-level-side-effect` to
   `no-top-level-side-effects`.
 - (`cec69d5`) _(Breaking)_ Report top-level IIFEs by default.